minor refactor and improved WG query
