+++ pom.xml
@@ -6,7 +6,7 @@
 
     <groupId>nl.aurorion.blockregen</groupId>
     <artifactId>BlockRegen</artifactId>
-    <version>3.5.8-b1</version>
+    <version>3.5.8-b2</version>
 
     <name>BlockRegen</name>
