+++ src/main/java/nl/aurorion/blockregen/BlockRegen.java
@@ -7,19 +7,19 @@ import com.sk89q.worldguard.bukkit.WorldGuardPlugin;
 import lombok.Getter;
 import me.ryanhamshire.GriefPrevention.GriefPrevention;
 import net.milkbowl.vault.economy.Economy;
-import nl.aurorion.blockregen.Commands.Commands;
-import nl.aurorion.blockregen.Configurations.Files;
-import nl.aurorion.blockregen.Events.BlockBreak;
-import nl.aurorion.blockregen.Events.DependencyEnable;
-import nl.aurorion.blockregen.Events.PlayerInteract;
-import nl.aurorion.blockregen.Events.PlayerJoin;
-import nl.aurorion.blockregen.Particles.ParticleUtil;
-import nl.aurorion.blockregen.System.ConsoleOutput;
-import nl.aurorion.blockregen.System.Getters;
-import nl.aurorion.blockregen.System.UpdateCheck;
-import nl.aurorion.blockregen.provider.JobsProvider;
-import nl.aurorion.blockregen.provider.WorldEditProvider;
-import nl.aurorion.blockregen.provider.WorldGuardProvider;
+import nl.aurorion.blockregen.commands.Commands;
+import nl.aurorion.blockregen.configuration.Files;
+import nl.aurorion.blockregen.listeners.BlockBreak;
+import nl.aurorion.blockregen.listeners.DependencyEnable;
+import nl.aurorion.blockregen.listeners.PlayerInteract;
+import nl.aurorion.blockregen.listeners.PlayerJoin;
+import nl.aurorion.blockregen.particles.ParticleUtil;
+import nl.aurorion.blockregen.system.ConsoleOutput;
+import nl.aurorion.blockregen.system.Getters;
+import nl.aurorion.blockregen.system.UpdateCheck;
+import nl.aurorion.blockregen.providers.JobsProvider;
+import nl.aurorion.blockregen.providers.WorldEditProvider;
+import nl.aurorion.blockregen.providers.WorldGuardProvider;
 import org.bukkit.ChatColor;
 import org.bukkit.Location;
 import org.bukkit.Material;
