+++ src/main/java/nl/aurorion/blockregen/configuration/ConfigFile.java
@@ -1,4 +1,4 @@
-package nl.aurorion.blockregen.Configurations;
+package nl.aurorion.blockregen.configuration;
 
 import lombok.Getter;
 import nl.aurorion.blockregen.BlockRegen;
