+++ src/main/java/nl/aurorion/blockregen/system/Amount.java
@@ -1,4 +1,4 @@
-package nl.aurorion.blockregen.System;
+package nl.aurorion.blockregen.system;
 
 import com.google.common.base.Strings;
 import lombok.Getter;
